Fix: Do not crash URL Auth is used with stream_auth and no credentials are given
This fixes a crash (NULL reference) in case URL Auth is used
and stream_auth is trigged with no credentials passed by the client.
Username and password is now set to empty strings and transmited to
the backend server this way.

See #2191 for more details and to keep track of the problem.

Closes: #2191, DEB#782120