Ensure VORBIS_COMMENT is attached for FLAC streams
Since commit c80fd692b2e3 ("Feature: Added basic FLAC
METADATA_BLOCK_HEADER parser") the VORBIS_COMMENT header is no longer
attached, resulting in a broken FLAC stream. Web browsers refuse to
play it, and libavformat warns about:
Headers mismatch for stream 0: expected 2 received 1.

This occurs because the previous implementation would set
codec->headers = 0 when 0xFF was found, which was after the header
had already been queued. Since it now detects the actual last header
(VORBIS_COMMENT), setting the headers field to 0 results in
format_ogg_attach_header() never being called for it.

It is now safe to remove if guard around it due to the aforementioned
change, thus resolving the issue. The headers field is still set to 0
which means that on the next call, after VORBIS_COMMENT, the code will
not be executed as expected.

Fixes: c80fd692b2e3 ("Feature: Added basic FLAC METADATA_BLOCK_HEADER parser")
Closes: #2493
Link: https://gitlab.xiph.org/xiph/icecast-server/-/issues/2493
Signed-off-by: Tasos Sahanidis <tasos@tasossah.com>